Closes #244. Most of the conversion had already been done. There could still be the  odd Int left.

+++ Util/srcGeom/TextGraphic.scala
@@ -48,7 +48,7 @@ object TextGraphic
   def xy(str: String, fontSize: Double = 24, xPosn: Double, yPosn: Double, colour: Colour = Black, align: TextAlign = CenAlign,
     baseLine: BaseLine = BaseLine.Middle) = new TextGraphic(str, fontSize, xPosn, yPosn, colour, align, baseLine)
 
-  def lines(strs: StrArr, fontSize: Int = 24, posn: Pt2 = Pt2Z, fontColour: Colour = Black, lineSpacing: Double = 1,
+  def lines(strs: StrArr, fontSize: Double = 24, posn: Pt2 = Pt2Z, fontColour: Colour = Black, lineSpacing: Double = 1,
     align: TextAlign = CenAlign, baseLine: BaseLine = BaseLine.Alphabetic): RArr[TextGraphic] =
   { val len = strs.length
     if(len == 0) RArr()
