+++ Util/srcShapes/Polygon.scala
@@ -268,20 +268,20 @@ trait Polygon extends Shape with BoundedElem with Approx[Double] with Pt2SeqSpec
   def fillDrawActive(fillColour: Colour, pointerID: AnyRef, lineWidth: Double, lineColour: Colour = Black): PolygonCompound =
     PolygonCompound(this, RArr(fillColour, DrawFacet(lineColour, lineWidth)), RArr(PolygonActive(this, pointerID)))
 
-  def fillDrawText(fillColour: Colour, str: String, fontSize: Int = 24, lineColour: Colour = Black, lineWidth: Double = 2.0): PolygonCompound =
+  def fillDrawText(fillColour: Colour, str: String, fontSize: Double = 24, lineColour: Colour = Black, lineWidth: Double = 2.0): PolygonCompound =
     PolygonCompound(this, RArr(fillColour, DrawFacet(lineColour, lineWidth)), RArr(TextGraphic(str, fontSize, cenDefault)))
 
   def parentFillText(pointerID: Any, fillColour: Colour, str: String, fontSize: Int = 10, textColour: Colour = Black, align: TextAlign = CenAlign):
   PolygonCompound = PolygonCompound(this, RArr(fillColour, TextFacet(str, textColour)), RArr())
 
-  def fillDrawTextActive(fillColour: Colour, pointerID: AnyRef, str: String, fontSize: Int = 24, lineWidth: Double, lineColour: Colour = Black,
+  def fillDrawTextActive(fillColour: Colour, pointerID: AnyRef, str: String, fontSize: Double = 24, lineWidth: Double, lineColour: Colour = Black,
     align: TextAlign = CenAlign): PolygonCompound = PolygonCompound(this, RArr(fillColour, DrawFacet(lineColour, lineWidth)),
     RArr(TextGraphic(str, fontSize, cenDefault, Black, align), PolygonActive(this, pointerID)))
 
-  def fillText(fillColour: Colour, str: String, fontSize: Int = 10, textColour: Colour = Black, layer: Int = 0): PolygonCompound =
+  def fillText(fillColour: Colour, str: String, fontSize: Double = 10, textColour: Colour = Black, layer: Int = 0): PolygonCompound =
     PolygonCompound(this, RArr(fillColour), RArr(TextGraphic(str, fontSize, cenDefault, textColour)))
 
-  def fillTextActive(fillColour: Colour, pointerEv: AnyRef, str: String, fontSize: Int = 24, fontColour: Colour = Black, align: TextAlign = CenAlign):
+  def fillTextActive(fillColour: Colour, pointerEv: AnyRef, str: String, fontSize: Double = 24, fontColour: Colour = Black, align: TextAlign = CenAlign):
     PolygonCompound = PolygonCompound(this, RArr(fillColour), RArr(PolygonActive(this, pointerEv), TextGraphic(str, fontSize, cenDefault, fontColour, align)))
 
   /** Insert vertex. */
